moved the Tag classes into tags namespace
